+++ lib/src/widgets/collections/collections.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/widgets.dart';
+
+typedef CollectionWidgetBuilder<T> = Widget Function(
+  BuildContext context,
+  T item,
+  int index,
+);